+++ src/includes/functions/.load.php
@@ -11,6 +11,7 @@
 /*
  * Load all functions.
  */
+require_once __DIR__.'/actions.php';
 require_once __DIR__.'/checks.php';
 require_once __DIR__.'/filters.php';
 require_once __DIR__.'/notices.php';